Revert "Moved management controllers into webcontroller package"
This reverts commit 926d76052704b535ffed5e4deaa3d9b689d05d2a.
