+++ core/src/main/java/greencity/controller/ManagementFactOfTheDayController.java
@@ -1,4 +1,4 @@
-package greencity.webcontroller;
+package greencity.controller;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
